Move components/raw -> components/common, components/cirrus/src -> components/common/src/cirrus
