+++ src/components/common/LICENSE
