+++ src/components/common/src/domain/firebase-auth-ui.tsx
