+++ src/components/common/src/util/lazy-component-creater.tsx
