--- src/lib/lazy-components/firebase-auth-ui.ts
@@ -1,6 +1,0 @@
-import { FirebaseAuthUI } from "@components/cirrus/domain/firebase-auth-ui";
-import { createLazyComponent } from "@components/cirrus/util/lazy-component-creater";
-
-export const createLazyFirebaseAuthUI = (context: FirebaseAuthUI.Context) => {
-  return createLazyComponent(() => import('@components/cirrus/domain/firebase-auth-ui'), 'FirebaseAuthUI', context)
-}
